Rename: 커뮤니티 Repository 위치 변경
Related to: #47
