+++ src/main/java/pulleydoreurae/careerquestbackend/community/repository/PostLikeRepository.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.common.community.repository;
+package pulleydoreurae.careerquestbackend.community.repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -8,8 +8,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
-import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
-import pulleydoreurae.careerquestbackend.common.community.domain.entity.PostLike;
+import pulleydoreurae.careerquestbackend.community.domain.entity.Post;
+import pulleydoreurae.careerquestbackend.community.domain.entity.PostLike;
 
 /**
  * 게시글 좋아요 Repository
