+++ src/main/java/pulleydoreurae/careerquestbackend/community/repository/PostViewCheckRepository.java
@@ -1,8 +1,8 @@
-package pulleydoreurae.careerquestbackend.common.community.repository;
+package pulleydoreurae.careerquestbackend.community.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import pulleydoreurae.careerquestbackend.common.community.domain.entity.PostViewCheck;
+import pulleydoreurae.careerquestbackend.community.domain.entity.PostViewCheck;
 
 /**
  * 조회수 중복을 막기 위해 사용되는 객체의 Repository
